+++ piethorn/math/converter/__init__.py
@@ -1,5 +1,5 @@
 """
-pythorn.math.converter
+piethorn.math.converter
 
 One-to-one Python port of the Java `cache` utility.